pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-sql-analyst-env"
version = "0.1.0"
description = "SQL Data Analysis Agent Environment for OpenEnv — e-commerce BI tasks"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "sql_analyst_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["sql_analyst_env", "sql_analyst_env.server"]
package-dir = { "sql_analyst_env" = ".", "sql_analyst_env.server" = "server" }
